+++ apps/docs/app/(web)/(routes)/[...slug]/page.tsx
@@ -11,6 +11,7 @@ import { SideMenu } from "@/components/navigation/side-menu/side-menu"
 import { PageBuilder } from "@/components/page-builder/page-builder"
 import pageLayoutStyles from "@/styles/page-layout.module.css"
 import { DraftTools } from "../_draftTools/DraftTools"
+import slugPageStyles from "./slugPage.module.css"
 
 async function fetchSlugPageData(slug: string) {
   return await sanityFetch({
@@ -74,7 +75,7 @@ export default async function SlugPage({ params }: { params: Promise<{ slug: str
     <>
       <div className={cn(pageLayoutStyles["page-layout"], pageLayoutStyles["page-spacing"])}>
         {rootSlug && <SideMenu slug={slug} rootSlug={rootSlug} />}
-        <main>
+        <main className={slugPageStyles["wrapper"]}>
           {showPageBuilder && <PageBuilder pageBuilder={pageBuilder} id={_id} type={_type} />}
 
           {showSubPageGrid && (