+++ apps/docs/components/page-builder/rich-text.tsx
@@ -1,4 +1,4 @@
-import { Heading, Ingress } from "@gyldendal/kobber-components/react"
+import { Heading, Lead } from "@gyldendal/kobber-components/react"
 import { CheckCircle, Remove } from "@gyldendal/kobber-icons/react"
 import {
   PortableText,
@@ -22,7 +22,7 @@ const components: Partial<PortableTextReactComponents> = {
     h2: ({ children, value }) => {
       const slug = parseChildrenToSlug(value.children)
       return (
-        <Heading id={slug} level="h2" element="title" size="medium" color-level="primary">
+        <Heading id={slug} level="h2" size="medium">
           {children}
         </Heading>
       )
@@ -30,14 +30,14 @@ const components: Partial<PortableTextReactComponents> = {
     h2Italic: ({ children, value }) => {
       const slug = parseChildrenToSlug(value.children)
       return (
-        <Heading id={slug} level="h2" element="display" size="small" color-level="secondary">
+        <Heading id={slug} level="h2" highlighted>
           {children}
         </Heading>
       )
     },
     h3: ({ children, value }) => {
       const slug = parseChildrenToSlug(value.children)
-      return <Ingress id={slug}>{children}</Ingress>
+      return <Lead id={slug}>{children}</Lead>
     },
     inline: ({ children }) => <p className={styles["inline"]}>{children}</p>,
   },