+++ apps/docs/eslint.config.js
@@ -1,6 +1,0 @@
-import { kobberEslintNextConfig } from "@gyldendal/kobber-eslint/next-js"
-
-/**@type {import("eslint").Linter.Config} */
-export default [
-  ...kobberEslintNextConfig,
-]