+++ apps/storybook-web-components/.storybook/kobberTheme.js
@@ -1,7 +1,8 @@
 import { create } from "storybook/theming";
 import { semantics } from "@gyldendal/kobber-base/themes/default/tokens.js";
 
-const background = semantics.color.identity.base["aubergine-1000"];
+const primary = semantics.color.identity.base["aubergine-1000"];
+const secondary = semantics.typography.color.brand["aubergine-25"];
 
 /** Options: https://storybook.js.org/docs/configure/user-interface/theming */
 export default create({
@@ -12,11 +13,9 @@ export default create({
   brandTitle: "Kobber Storybook",
   brandImage: "https://dam-prod.gyldendaldigital.no/tenants/edu/file/D_8MwlmJKANA6W4OIRWa7k/*/logo.svg",
 
-  colorSecondary: semantics.typography.color.bright,
+  appBg: primary,
+  appContentBg: primary,
+  appPreviewBg: secondary,
 
-  appBg: background,
-  appContentBg: background,
-  appPreviewBg: semantics.color.identity.base["aubergine-25"],
-
-  barBg: background,
+  barBg: primary,
 });