+++ packages/kobber-components/src/text/Text.stories.ts
@@ -2,14 +2,19 @@ import type { Meta, StoryObj } from "@storybook/web-components-vite";
 import { html } from "lit/static-html.js";
 import "./text-wrapper/TextWrapper";
 import "./heading/Heading";
-import "./ingress/Ingress";
+import "./lead/Lead";
 import "./text-link/TextLink";
-import { headingElements, headingSizes, headingColorLevels } from "./heading/Heading.core";
+import {
+  headingColors,
+  headingSizes,
+  headingColorVariants,
+  headingFonts,
+} from "./heading/Heading.core";
 import "@gyldendal/kobber-icons/web-components";
 import { init as initComponents } from "../base/init";
 import { init as initIcons } from "@gyldendal/kobber-icons/init";
 import { getPrintedState, linkStates } from "../story/linkStates";
-import { ingressSizes } from "./ingress/Ingress.core";
+import { leadColors, leadColorVariants } from "./lead/Lead.core";
 
 initComponents();
 initIcons();
@@ -36,11 +41,11 @@ export const All: Story = {
         <em>Gyldendals designsystem</em>
       </kobber-heading>
 
-      <kobber-ingress>
+      <kobber-lead>
         Phosfluorescently innovate real-time experiences vis-a-vis unique opportunities. Interactively disintermediate
         sustainable niches before long-term <em>high-impact</em> resources. Interactively deliver 2.0 infomediaries via
         timely.
-      </kobber-ingress>
+      </kobber-lead>
 
       <p>
         Rapidiously strategize integrated intellectual capital vis-a-vis
@@ -116,44 +121,48 @@ export const Heading: Story = {
     h1: false,
   },
   render: args => {
-    const text = (textValue: string) => (args.highlighted ? html`<em>${textValue}</em>` : textValue);
+    const text = (textValue: string) =>
+      args.highlighted ? html`<em>${textValue}</em>` : textValue;
 
     return html`
       <kobber-text-wrapper class="kobber-text-wrapper">
-        ${headingElements.map(
-          element => html`
+        ${headingColors.map(
+          color => html`
             ${headingSizes.map(
               size =>
-                html` <div
-                  style="
-                        display: grid;
-                        grid-template-columns: 1fr 1fr;
-                        align-items: center;
-                        grid-template-areas: 
-                          'element-and-size .'
-                          'primary sample-primary'
-                          'secondary sample-secondary'
-                          'reading sample-reading'
-                          'ui sample-ui';
-                        border: 1px solid;
-                        padding: 1em;"
-                >
-                  <em style="grid-area: element-and-size;">${element} - ${size}</em>
-                  ${headingColorLevels.map(
-                    colorVariant => html`
-                      <p style="grid-area: ${colorVariant};">${colorVariant}</p>
-                      <kobber-heading
-                        level="${args.h1 ? "h1" : "h2"}"
-                        element="${element}"
-                        color-variant="${colorVariant}"
-                        size="${size}"
-                        style="grid-area: sample-${colorVariant};"
-                      >
-                        ${text(args.text || "Heading")}
-                      </kobber-heading>
-                    `,
-                  )}
-                </div>`,
+                html` <div></div>
+            <em style="font-size: 1.2em;">${color} - ${size}</em>
+            <div
+            style="
+              display: grid;
+              border: 1px solid;
+              padding: 1em;"
+            >
+            ${headingColorVariants.map(colorVariant =>
+              headingFonts.map(
+                font => html`
+                  <div
+                    style="margin-bottom: 0.5rem; background-color: ${
+                      colorVariant === "tone-b" ? "darkgray" : "transparent"
+                    };"
+                  >
+                    <p style="">${colorVariant} ${font}</p>
+                    <kobber-heading
+                      highlighted
+                      level="${args.h1 ? "h1" : "h2"}"
+                      size="${size}"
+                      font="${font}"
+                      color="${color}"
+                      colorVariant="${colorVariant}"
+                    >
+                      ${text(args.text || "Heading")}
+                    </kobber-heading>
+                  </div>
+                `,
+              ),
+            )}
+            </div>
+          </div>`,
             )}
           `,
         )}
@@ -163,25 +172,57 @@ export const Heading: Story = {
 };
 
 /**
+ * 
+                    html`
+                      <p style="grid-area: ${colorVariant};">${colorVariant}</p>
+                      <kobber-heading
+                        level="${args.h1 ? "h1" : "h2"}"
+                        size="${size}"
+                        font="${"brand"}"
+                        color="${color}"
+                        colorVariant="${colorVariant}"
+                        style="grid-area: sample-${colorVariant};"
+                      >
+                        ${text(args.text || "Heading")}
+                      </kobber-heading>
+                      <p style="grid-area: ${font};">${font}</p>
+                      <kobber-heading
+                        level="${args.h1 ? "h1" : "h2"}"
+                        size="${size}"
+                        font="${font}"
+                        color="${color}"
+                        colorVariant="${colorVariant}"
+                        style="grid-area: sample-${colorVariant};"
+                      >
+                        ${text(args.text || "Heading")}
+                      </kobber-heading>
+                    `,
+                  )
  * Bruker farge fra "component.article", og typografi fra "title medium".
  */
-export const Ingress: Story = {
+export const Lead: Story = {
   argTypes: {
     text: {
       control: "text",
     },
-    size: {
-      options: ingressSizes,
-      control: { type: "inline-radio" },
+    color: {
+      control: "inline-radio",
+      options: leadColors,
+    },
+    colorVariant: {
+      control: "inline-radio",
+      options: leadColorVariants,
+      name: "color-variant",
     },
   },
   args: {
-    text: "Kobber er Gyldendals verktøykasse for design- og merkevare. Det er et designsystem bestående av gjenbrukbare, fleksible ressurser slik som digitale komponenter, malverk, retningslinjer og kode. Samtidig tydeliggjør det vår merkevarestrategi, våre felles verdier og de opplevelsene vi har som mål å tilby våre sluttbrukere.",
-    size: "small",
+    text: "Lead er en ingress som brukes som en kort innledningstekst som oppsummerer eller introduserer innholdet.",
+    color: leadColors[0],
+    colorVariant: leadColorVariants[0],
   },
-  render: args => {
+  render: (args) => {
     return html`<div style="max-width: 600px;">
-      <kobber-ingress size="${args.size}"> ${args.text} </kobber-ingress>
+      <kobber-lead color="${args.color}" color-variant="${args.colorVariant}"> ${args.text} </kobber-lead>
     </div>`;
   },
 };