+++ packages/kobber-audio-recorder/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gyldendal/kobber-audio-recorder
 
+## 0.6.5
+
+### Patch Changes
+
+- Add new icons, remove kobber-eslint and kobber-prettier
+
+## 0.6.4
+
+### Patch Changes
+
+- Add new icons
+
+## 0.6.3
+
+### Patch Changes
+
+- kobber-audio-recorder macos-safari-bugfix
+
 ## 0.6.2
 
 ### Patch Changes